server/requirements.txt
fastapi
pydantic
uvicorn
python-box
boto3
google-cloud-storage